activate new texture model in light.hh, and therefore add -DNDEBUG in CMake
Probably because the 'intersection' vector provided by illumination() is not
always in the intersected Shape, assert()ions in computeColorFromTexture() are
triggered (experimented only with Triangle). However, on tested scenes
(without textures), the resulting image is the same as with the old model
(retrieving the color of the texture at point (0,0)).

Going to unit-test Triangle::containsPoint() (used in the
Triangle::computeColorFromTexture()'s assert) to ensure the problem comes from
the imprecision of 'intersection'.